Fix SSF-125
