+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/server/NetUtils.java
@@ -20,10 +20,7 @@
 package org.sonar.css.plugin.server;
 
 import java.io.IOException;
-import java.net.InetSocketAddress;
 import java.net.ServerSocket;
-import java.net.Socket;
-import java.net.SocketAddress;
 
 public class NetUtils {
 
@@ -36,29 +33,4 @@ public class NetUtils {
     }
   }
 
-  public static boolean waitServerToStart(String host, int port, int timeoutMs) {
-    int sleepStep = 20;
-    SocketAddress address = new InetSocketAddress(host, port);
-    long start = System.currentTimeMillis();
-    try {
-      while (!serverListening(address)) {
-        if (System.currentTimeMillis() - start > timeoutMs) {
-          return false;
-        }
-        Thread.sleep(sleepStep);
-      }
-    } catch (InterruptedException e) {
-      Thread.currentThread().interrupt();
-    }
-    return true;
-  }
-
-  private static boolean serverListening(SocketAddress address) {
-    try (Socket s = new Socket()) {
-      s.connect(address, 1);
-      return true;
-    } catch (Exception e) {
-      return false;
-    }
-  }
 }
